Update actor with attack sets
